requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
get==2019.4.13
post==2019.4.13
public==2019.4.13
query-string==2019.4.13
request==2019.4.13
soupsieve==2.0.1
